generator/requirements.txt
et-xmlfile==1.1.0
fpdf==1.7.2
lxml==5.2.2
numpy==2.0.1
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
